Moved ActivityTask into worker
