+++ src/main/java/com/uber/cadence/internal/worker/SynchronousActivityTaskPoller.java
@@ -27,7 +27,6 @@ import com.uber.cadence.TaskList;
 import com.uber.cadence.WorkflowExecution;
 import com.uber.cadence.WorkflowService;
 import com.uber.cadence.internal.common.SynchronousRetrier;
-import com.uber.cadence.internal.sync.ActivityExecutionException;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.thrift.TException;
@@ -193,7 +192,7 @@ class SynchronousActivityTaskPoller implements TaskPoller {
             if (activityImplementation == null) {
                 throw new IllegalArgumentException("Unknown activity type: " + activityType);
             }
-            output = activityImplementation.execute(service, domain, task);
+            output = activityImplementation.execute(service, domain, new ActivityTaskImpl(task));
             if (!activityImplementation.getExecutionOptions().isDoNotCompleteOnReturn()) {
                 respondActivityTaskCompletedWithRetry(task.getTaskToken(), output);
             }
@@ -211,7 +210,7 @@ class SynchronousActivityTaskPoller implements TaskPoller {
                         + ", activityId=" + task.getActivityId(), e);
             }
             if (!(e instanceof ActivityExecutionException)) {
-                e = activityImplementationFactory.serializeUnexpectedFailure(task, e);
+                e = activityImplementationFactory.serializeUnexpectedFailure(new ActivityTaskImpl(task), e);
             }
             ActivityExecutionException executionException = (ActivityExecutionException) e;
             respondActivityTaskFailedWithRetry(task.getTaskToken(), executionException.getReason(),
